Leave only package loading in JuliaInterpreter.jl
